+++ lib/latex-provider.js
@@ -2,7 +2,9 @@ const fs = require("fs");       // inbuilt nodejs file system tool
 const path = require("path");   // inbuilt nodejs file path tool
 const _ = require("lodash");    // miscellaneous useful functions
 const fz = require("fuzzaldrin-plus"); // for sorting of array entries based on some value
-const { parseScopeString, scopesMatch } = require("./scope-tools");
+const { parseScopeString, scopesMatch, mergeCompletionObjects } = require("./scope-tools");
+const { citationCompletions } = require("./bib-tools");
+const { CacheManager } = require("./cache-manager");
 const DEFAULT_COMPLETIONS = require("./resources/completions.json");
 var USER_COMPLETIONS = {};
 var COMPLETIONS = {};
@@ -22,7 +24,22 @@ module.exports = { // deliberately chosen `undefined` for config properties prob
   citationControlWord: "autocite",
   cache: undefined, // set up in this.loadProperties
 
-  loadProperties(serialized) {
+  serialize() {
+    return {
+      cache: this.cache.serialize()
+    };
+  },
+
+  deserialize(state) {
+      this.cache = new CacheManager();
+      if (state && state.cache) {
+        this.cache.deserialize(state.cache);
+      }
+  },
+
+  loadProperties() {
+    if (!this.cache) { this.cache = new CacheManager(); }
+
     // get config values
     this.prefixRegex =  new RegExp(atom.config.get("autocomplete-latex.commandCompletionRegex"));
     this.citationPrefixRegex =  new RegExp(atom.config.get("autocomplete-latex.citationCompletionRegex"));
@@ -30,10 +47,8 @@ module.exports = { // deliberately chosen `undefined` for config properties prob
     this.minPrefixLength = atom.config.get("autocomplete-latex.minPrefixLength");
     this.citations = atom.config.get("autocomplete-latex.enableCitationCompletions");
     this.excludeLowerPriority = !atom.config.get("autocomplete-latex.enableBuiltinProvider");
-    let userCiteWord = atom.config.get("autocomplete-latex.citationControlWord");
-    if (userCiteWord) { this.citationControlWord = userCiteWord; }
 
-    // label all default completiosn accordingly
+    // label all default completions accordingly
     for (let scope in DEFAULT_COMPLETIONS) {
       for (let type in DEFAULT_COMPLETIONS[scope]) {
         for (let i = 0; i < DEFAULT_COMPLETIONS[scope][type].length; i++) {
@@ -49,59 +64,29 @@ module.exports = { // deliberately chosen `undefined` for config properties prob
     }
     if (fs.existsSync(pathToUserCompletions)) { // if it can find the file
       try {
-        fs.stat(pathToUserCompletions, (err, stats) => {
-          console.log(stats.mtime);
-        });
-        USER_COMPLETIONS = require(pathToUserCompletions);
+        let metaData = JSON.parse(JSON.stringify(fs.statSync(pathToUserCompletions))); // to make the data serialisable
+        if (!this.cache.hasCurrentCache(pathToUserCompletions, metaData)) {
+          USER_COMPLETIONS = require(pathToUserCompletions);
+          this.cache.empty(); // clears out old completions
+          this.cache.setCache(pathToUserCompletions, USER_COMPLETIONS, metaData);
+        } else {
+          // NOTE: caching here is kind of pointless, given it'll be reading the cached data from a file anyway.
+          // It will be better when the merged object is cached, so the spaghetti code can be bypassed.
+          USER_COMPLETIONS = this.cache.getCache(pathToUserCompletions);
+        }
       }
       catch (err) {
-        atom.notifications.addError(`\`autocomplete-latex\`\nFailed to read user completions.\n- ${err}`, {dismissable: true});
+        atom.notifications.addWarning(`\`autocomplete-latex\`\nFailed to read user completions.\n- ${err}`, {dismissable: true});
         USER_COMPLETIONS = {};
       }
     }
 
     // merges user + default completions according to settings
     if (atom.config.get("autocomplete-latex.enableDefaultCompletions")) {
-      COMPLETIONS = this.mergeCompletionObjects(USER_COMPLETIONS, DEFAULT_COMPLETIONS);
+      COMPLETIONS = mergeCompletionObjects(USER_COMPLETIONS, DEFAULT_COMPLETIONS);
     } else {
       COMPLETIONS = USER_COMPLETIONS;
     }
-
-    if (serialized.cachedCompletions) {
-      this.cache = serialized.cachedCompletions;
-    } else {
-      this.cache = new Map();
-    }
-  },
-
-  mergeCompletionObjects(first, second) { // first will take priority in any duplicates
-    // TODO: Refactor bad code and relocate to a utilities file (maybe `scope-helpers`?)
-    let merged = JSON.parse(JSON.stringify(first)); // make an unlinked copy
-    let clashes = [];
-    for (let scopeKey in second) {
-      if (scopeKey in first) {
-        for (let typeKey in second[scopeKey]) {
-          if (typeKey in first[scopeKey]) {
-            for (let i = 0; i < second[scopeKey][typeKey].length; i++) {
-              let secondDisplay = second[scopeKey][typeKey][i].displayText || second[scopeKey][typeKey][i].prefix;
-              for (let j = 0; j < first[scopeKey][typeKey].length; j++) {
-                let firstDisplay = first[scopeKey][typeKey][j].displayText || first[scopeKey][typeKey][j].prefix;
-                if (firstDisplay === secondDisplay) {
-                  break;
-                } else if (j === first[scopeKey][typeKey].length - 1) {
-                  merged[scopeKey][typeKey].push(second[scopeKey][typeKey][i]);
-                }
-              }
-            }
-          } else {
-            merged[scopeKey][typeKey] = second[scopeKey][typeKey];
-          }
-        }
-      } else {
-        merged[scopeKey] = second[scopeKey];
-      }
-    }
-    return merged;
   },
 
   getSuggestions({editor, bufferPosition, scopeDescriptor}) { // autocalled by `autocomplete-plus`
@@ -110,13 +95,12 @@ module.exports = { // deliberately chosen `undefined` for config properties prob
     if (this.prefix.length < this.minPrefixLength) { return; }
 
     if (this.prefix[0] === "@") {
-      return this.citationCompletions(this.prefix, editor);
+      return citationCompletions(this.prefix, editor);
     }
 
     var completions = [];
-
-    if (this.cache.has(scopeDescriptor)) {
-      completions = this.cache.get(scopeDescriptor);
+    if (this.cache.hasCache(scopeDescriptor.scopes.join("|"))) { // join character is arbitrary; may need to change if | is used in future scope rules
+      completions = this.cache.getCache(scopeDescriptor.scopes.join("|"));
     } else {
       for (let scope in COMPLETIONS) {
         if (!scopesMatch(parseScopeString(scope), scopeDescriptor.scopes)) { continue; }
@@ -128,7 +112,7 @@ module.exports = { // deliberately chosen `undefined` for config properties prob
               text: suggestion.text, // OR
               snippet: suggestion.snippet,
               displayText: suggestion.displayText || suggestion.prefix, // allows explicit setting of what UI shows. Defaults to prefix if not provided (prefix is deprecated)
-              replacementPrefix: suggestion.replacementPrefix || this.prefix, // is used to delete the existing characters
+              // replacementPrefix: suggestion.replacementPrefix || this.prefix, // is used to delete the existing characters. Useless with caching.
               type: suggestion.type || compType,
               leftLabel: suggestion.leftLabel,
               leftLabelHTML: suggestion.leftlabelHTML,
@@ -144,93 +128,17 @@ module.exports = { // deliberately chosen `undefined` for config properties prob
           }
         }
       }
-      this.cache.set(scopeDescriptor, completions);
+      this.cache.setCache(scopeDescriptor.scopes.join("|"), completions);
     }
 
+    // Short list the candidates
     completions = fz.filter(completions, this.prefix, {key:"displayText"});
-    completions = _.uniqBy(completions, (e) => { return e.displayText });
+    completions = _.uniqBy(completions, (e) => { return e.displayText; });
 
-    return completions;
-  },
-
-  citationCompletions(prefix, editor) {
-    let bibFilePath = this.findBibFile(editor);
-    if (!bibFilePath) { return; }
-
-    let completions;
-    let data;
-    try {
-      data = fs.readFileSync(bibFilePath, "utf-8")
-    } catch (err) {
-      console.warn(`autocomplete-latex could not find bib file:\n${err}`);
-      return false;
-    }
-    completions =  this.parseBibFile(data, prefix);
-    completions = fz.filter(completions, prefix, {key:"displayText", allowErrors:true}); // allowErrors allows any chaarcters
-    return completions;
-  },
-
-  findBibFile(editor) {
-    let fileContents = editor.getText();
-    let magicBibPath = fileContents.match(/% !T[eE]X bib =\s*(.*)/);
-    if (magicBibPath) { // if bib path explicity set, go with that one
-      return path.resolve(editor.getDirectoryPath(), magicBibPath[1]);
-    }
-
-    let bibPath = fileContents.match(/\\addbibresource\{(.*?)\}/);
-    if (bibPath) {
-      bibPath[1] = bibPath[1].trim();
-      return path.resolve(editor.getDirectoryPath(), bibPath[1]);
-    }
-
-    let rootFile = fileContents.match(/% !T[eE]X root =\s*(.*)/);
-    if (!rootFile) { return false; }
-
-    rootFile = path.resolve(editor.getDirectoryPath(), rootFile[1]);
-    try {
-      fileContents = fs.readFileSync(rootFile, "utf-8");
-    } catch (err) {
-      // Fails silently.
-      console.warn(`autocomplete-latex could not find root file:\n${err}`);
-      return false;
-    }
-
-    let rootFilePath = path.dirname(rootFile);
-
-    magicBibPath = fileContents.match(/% !T[eE]X bib =\s*(.*)/);
-    if (magicBibPath) { // if bib path explicity set, go with that one
-      return path.resolve(rootFilePath, magicBibPath[1]);
-    }
-
-    bibPath = fileContents.match(/\\addbibresource\{(.*?)\}/);
-    if (bibPath) {
-      return path.resolve(rootFilePath, bibPath[1]);
-    }
-
-    // Could also look for \input{} files, as they may contain labels/bib paths
-
-    return false;
-  },
-
-  parseBibFile(file, prefix) {
-    file = file.substring(file.search("@") + 1); // Start just after the first @ sign (ignores any earlier text)
-    let entriesArray = file.split(/\s*@\s*/); // Split by the @ sign into separate strings. Needs work, as contents could have @ in it
-    let completions = [];
-    for (let i = 0; i < entriesArray.length; i++) {
-      let suggestion = {};
-      let entry = entriesArray[i];
-
-      let type = entry.match(/^\s*(.*?)\s*\{/);
-      if (!type) { continue; }
-      suggestion.type = type[1];
-
-      let displayText = entry.match(/\{\s*(.*?)\s*,/);
-      if (!displayText) { continue; }
-      suggestion.displayText = displayText[1];
-      suggestion.snippet = `\\\\${this.citationControlWord}\\{${suggestion.displayText}\\}$1`;
-      suggestion.replacementPrefix = prefix;
-
-      completions.push(suggestion);
+    // Fix the replacementPrefix for all that's left.
+    // Needed, as the cached values have the wrong replacement prefix.
+    for (let i = 0; i < completions.length; i++) {
+      completions[i].replacementPrefix = this.prefix;
     }
     return completions;
   },
@@ -245,28 +153,5 @@ module.exports = { // deliberately chosen `undefined` for config properties prob
     } else {
       return "";
     }
-  },
-
-  // reference for available types & their appearance
-  inbuiltTypes() {
-    return [
-      {text:"builtin", type:"builtin"},
-      {text:"class", type:"class"},
-      {text:"constant", type:"constant"},
-      {text:"function", type:"function"},
-      {text:"import", type:"import"},
-      {text:"keyword", type:"keyword"},
-      {text:"method", type:"method"},
-      {text:"module", type:"module"},
-      {text:"mixin", type:"mixin"},
-      {text:"package", type:"package"},
-      {text:"property", type:"property"},
-      {text:"require", type:"require"},
-      {text:"snippet", type:"snippet"},
-      {text:"tag", type:"tag"},
-      {text:"type", type:"type"},
-      {text:"value", type:"value"},
-      {text:"variable", type:"variable"}
-    ];
   }
 };
